requirements.txt
torch
transformers
faiss-cpu
numpy
PyPDF2